Lecture10/phdays/dao_2/pyproject.toml
[tool.poetry]
name = "dao-2"
version = "0.1.0"
description = ""
authors = ["Sarkoxed <sarkoxed2013@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
web3 = "^6.19.0"
py-solc-x = "^2.0.3"
fastecdsa = "^2.3.2"
safe-pysha3 = "^1.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
